requirements.txt
fastapi
uvicorn
pyyaml
click
requests
pytest>=7.0.0
pytest-cov>=4.0.0
httpx>=0.24.0
